+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 4.4.0 ##
+
+1.  UI Toolkit: `Table` now support showing `Button` & `ShowInInspector`. (Note: Any kind of `Layout` will be ignored)
+2.  UI Toolkit: Fix `NoLabel` didn't work when inside a horizontal layout
+3.  UI Toolkit: Fix `Table` foldout out of area. If a struct is a field, that field will remove the `foldout`
+
 ## 4.3.3 ##
 
 UI Toolkit: Some unity built-in class type has some very weird behavior. Fallback to default drawing flow instead [#200](https://github.com/TylerTemp/SaintsField/issues/200)